Refactor project structure to move configs under core.
Moved configuration and exception files from the 'configs' directory to a new 'core' directory for better organization. Updated all relevant imports throughout the project to reflect this change.
