+++ backend/app/databases/db.py
@@ -4,7 +4,7 @@ from typing import AsyncGenerator
 from sqlalchemy.ext.asyncio import create_async_engine, async_sessionmaker, AsyncAttrs, AsyncSession
 from sqlalchemy.orm import DeclarativeBase
 
-from app.configs.config import config
+from app.core.config import config
 
 logging.getLogger("sqlalchemy.engine").setLevel(logging.INFO)
